Move rule and test to new file respectively - review comment
