+++ sql/core/src/test/scala/org/apache/spark/sql/UDFSuite.scala
@@ -20,8 +20,7 @@ package org.apache.spark.sql
 import java.math.BigDecimal
 
 import org.apache.spark.sql.api.java._
-import org.apache.spark.sql.catalyst.expressions.ScalaUDF
-import org.apache.spark.sql.catalyst.plans.logical.{LogicalPlan, Project}
+import org.apache.spark.sql.catalyst.plans.logical.Project
 import org.apache.spark.sql.execution.QueryExecution
 import org.apache.spark.sql.execution.columnar.InMemoryRelation
 import org.apache.spark.sql.execution.command.{CreateDataSourceTableAsSelectCommand, ExplainCommand}
@@ -150,41 +149,6 @@ class UDFSuite extends QueryTest with SharedSQLContext {
   test("TwoArgument UDF") {
     spark.udf.register("strLenScala", (_: String).length + (_: Int))
     assert(sql("SELECT strLenScala('test', 1)").head().getInt(0) === 5)
-  }
-
-  test("Deterministic and literal UDF optimization") {
-    def udfNodesCount(plan: LogicalPlan): Int = {
-      plan.expressions.head.children.collect({
-        case f: ScalaUDF => f
-      }).length
-    }
-
-    withSQLConf(SQLConf.DETERMINISTIC_UDF_FOLD_ENABLED.key -> "true") {
-      // Non deterministic
-      val foo = udf(() => Math.random())
-      spark.udf.register("random0", foo.asNondeterministic())
-      val plan = sql("SELECT random0()").queryExecution.optimizedPlan
-      assert(udfNodesCount(plan) == 1)
-
-      // udf is deterministic and args are literal
-      val foo2 = udf((x: String, i: Int) => x.length + i)
-      spark.udf.register("mystrlen", foo2)
-      assert(foo2.deterministic)
-      assert(sql("SELECT mystrlen('abc', 1)").head().getInt(0) == 4)
-      val plan2 = sql("SELECT mystrlen('abc', 1)").queryExecution.optimizedPlan
-      assert(udfNodesCount(plan2) == 0)
-      val plan3 = sql("SELECT mystrlen('abc', mystrlen('c', 1))").queryExecution.optimizedPlan
-      assert(udfNodesCount(plan3) == 0)
-
-      // udf is deterministic and args are not literal
-      withTempView("temp1") {
-        val df = sparkContext.parallelize(
-          (1 to 10).map(i => i.toString)).toDF("i1")
-        df.createOrReplaceTempView("temp1")
-        val plan = sql("SELECT mystrlen(i1, 1) FROM temp1").queryExecution.optimizedPlan
-        assert(udfNodesCount(plan) == 1)
-      }
-    }
   }
 
   test("UDF in a WHERE") {
